AP-2019 rename docker task to prevent name confusion with docker module import
