+++ tasks/__init__.py
@@ -12,7 +12,7 @@ from . import (
     cluster_agent_cloudfoundry,
     components,
     customaction,
-    docker,
+    docker_tasks,
     dogstatsd,
     epforwarder,
     github,
@@ -112,7 +112,7 @@ ns.add_collection(components)
 ns.add_collection(customaction)
 ns.add_collection(bench)
 ns.add_collection(trace_agent)
-ns.add_collection(docker)
+ns.add_collection(docker_tasks, "docker")
 ns.add_collection(dogstatsd)
 ns.add_collection(epforwarder)
 ns.add_collection(msi)
